// src/lib/server/type.d.ts
type CountryInfo = {
    name: string;
    capital: string;
    population: number;
    region: string;
    nativeName: string;
    subRegion: string;
    flagImgSrc: string;
    domain: string;
    currencies: string[];
    languages: string[];
    borderCountries: string[];
};

export type CountryInfo = CountryInfo;